move offerEqual to utils
