+++ core/zoe/utils.js
@@ -9,4 +9,24 @@ const transpose = matrix =>
     [],
   );
 
-export { allTrue, anyTrue, transpose };
+const ruleEqual = (leftRule, rightRule) => leftRule.rule === rightRule.rule;
+
+const amountEqual = (assay, leftRule, rightRule) =>
+  assay.equals(leftRule.amount, rightRule.amount);
+
+const offerEqual = (assays, leftOffer, rightOffer) => {
+  const isLengthEqual = leftOffer.length === rightOffer.length;
+  if (!isLengthEqual) {
+    return false;
+  }
+  return leftOffer
+    .map((leftRule, i) => {
+      return (
+        ruleEqual(leftRule, rightOffer[i]) &&
+        amountEqual(assays[i], leftRule, rightOffer[i])
+      );
+    })
+    .reduce(allTrue);
+};
+
+export { allTrue, anyTrue, transpose, offerEqual };
